requirements.txt
pyserial
firebase-admin
Flask
Flask-Cors
google-generativeai